Move OrganizationInvitation into github.NamedUser to avoid cyclic imports
