+++ github/NamedUser.py
@@ -691,3 +691,47 @@ class NamedUser(github.GithubObject.CompletableGithubObject):
             self._url = self._makeStringAttribute(attributes["url"])
         if "user_view_type" in attributes:  # pragma no branch
             self._user_view_type = self._makeStringAttribute(attributes["user_view_type"])
+
+# A better place would be github.OrganizationInvitation.OrganizationInvitation
+# but that causes an import cycle. This is a specialization of NamedUser any way.
+class OrganizationInvitation(NamedUser):
+    """
+    This class represents OrganizationInvitation.
+
+    The reference can be found here
+    https://docs.github.com/en/rest/orgs/members
+
+    The OpenAPI schema can be found at
+    - /components/schemas/organization-invitation
+
+    """
+
+    def _initAttributes(self) -> None:
+        super()._initAttributes()
+        self._failed_at: Attribute[str] = NotSet
+        self._failed_reason: Attribute[str] = NotSet
+        self._invitation_source: Attribute[str] = NotSet
+
+    def __repr__(self) -> str:
+        return self.get__repr__({"id": self._id.value})
+
+    @property
+    def failed_at(self) -> str:
+        return self._failed_at.value
+
+    @property
+    def failed_reason(self) -> str:
+        return self._failed_reason.value
+
+    @property
+    def invitation_source(self) -> str:
+        return self._invitation_source.value
+
+    def _useAttributes(self, attributes: dict[str, Any]) -> None:
+        super()._useAttributes(attributes)
+        if "failed_at" in attributes:  # pragma no branch
+            self._failed_at = self._makeStringAttribute(attributes["failed_at"])
+        if "failed_reason" in attributes:  # pragma no branch
+            self._failed_reason = self._makeStringAttribute(attributes["failed_reason"])
+        if "invitation_source" in attributes:  # pragma no branch
+            self._invitation_source = self._makeStringAttribute(attributes["invitation_source"])
